+++ tests/testthat/test-convert_orthologs.R
@@ -1,5 +1,7 @@
 test_that("convert_orthologs works", {
+    
     data("exp_mouse")
+    method <- "homologene"
     #### Prepare different formats ####
     exp_mouse_smat <- exp_mouse
     exp_mouse_dmat <- as.matrix(exp_mouse)
@@ -21,7 +23,8 @@ test_that("convert_orthologs works", {
         input_species = "mouse",
         gene_input = "rownames",
         gene_output = "rownames",
-        as_sparse = TRUE
+        as_sparse = TRUE,
+        method = method
     )
     testthat::expect_equal(methods::is(gene_smat1, "sparseMatrix"), TRUE)
     testthat::expect_equal(has_gene_cols(gene_smat1), 0)
@@ -30,7 +33,8 @@ test_that("convert_orthologs works", {
         gene_df = exp_mouse_dmat,
         input_species = "mouse",
         gene_input = "rownames",
-        gene_output = "rownames"
+        gene_output = "rownames",
+        method = method
     )
     testthat::expect_equal(methods::is(gene_dmat1, "matrix"), TRUE)
     testthat::expect_equal(has_gene_cols(gene_dmat1), 0)
@@ -39,7 +43,8 @@ test_that("convert_orthologs works", {
         gene_df = exp_mouse_dmat,
         input_species = "mouse",
         gene_input = "rownames",
-        gene_output = "columns"
+        gene_output = "columns",
+        method = method
     )
     testthat::expect_equal(methods::is(gene_dmatc1, "data.frame"), TRUE)
     testthat::expect_equal(has_gene_cols(gene_dmatc1), 2)
@@ -48,7 +53,8 @@ test_that("convert_orthologs works", {
         gene_df = exp_mouse_df,
         input_species = "mouse",
         gene_input = "rownames",
-        gene_output = "columns"
+        gene_output = "columns",
+        method = method
     )
     testthat::expect_equal(methods::is(gene_df1, "data.frame"), TRUE)
     testthat::expect_equal(has_gene_cols(gene_df1), 2)
@@ -59,11 +65,11 @@ test_that("convert_orthologs works", {
             gene_df = exp_mouse_df,
             input_species = "mouse",
             gene_input = "rownamesTYPO",
-            gene_output = "columns"
+            gene_output = "columns",
+            method = method
         )
     )
-
-
+    
     #### Test gene standardisation and sort rows ####
     gene_df_std <- convert_orthologs(
         gene_df = exp_mouse_df,
@@ -73,7 +79,8 @@ test_that("convert_orthologs works", {
         standardise_genes = TRUE,
         sort_rows = TRUE,
         drop_nonorths = FALSE,
-        non121_strategy = "kbs"
+        non121_strategy = "kbs",
+        method = method
     )
     testthat::expect_equal(methods::is(gene_df_std, "data.frame"), TRUE)
     testthat::expect_equal(has_gene_cols(gene_df_std), 3)
@@ -84,7 +91,8 @@ test_that("convert_orthologs works", {
         gene_df = gene_df_std,
         input_species = "mouse",
         gene_input = "input_gene",
-        gene_output = "columns"
+        gene_output = "columns",
+        method = method
     )
     testthat::expect_equal(methods::is(gene_df_converted, "data.frame"), TRUE)
     testthat::expect_equal(has_gene_cols(gene_df_converted), 3)
@@ -94,7 +102,8 @@ test_that("convert_orthologs works", {
     gene_dict <- convert_orthologs(
         gene_df = exp_mouse_df,
         input_species = "mouse",
-        gene_output = "dict"
+        gene_output = "dict",
+        method = method
     )
     testthat::expect_equal(methods::is(gene_dict, "character"), TRUE)
     testthat::expect_equal(methods::is(names(gene_dict), "character"), TRUE)
@@ -104,13 +113,40 @@ test_that("convert_orthologs works", {
     gene_dict_rev <- convert_orthologs(
         gene_df = rownames(exp_mouse_df)[seq(1, 100)],
         input_species = "mouse",
-        gene_output = "dict_rev"
+        gene_output = "dict_rev",
+        method = method
     )
     testthat::expect_equal(methods::is(gene_dict_rev, "character"), TRUE)
     testthat::expect_equal(methods::is(names(gene_dict_rev), "character"), TRUE)
     testthat::expect_equal(methods::is(unname(gene_dict_rev), "character"), TRUE)
     testthat::expect_equal(methods::is(gene_dict_rev, "vector"), TRUE)
-
+    
+    # vector input - dataframe rownames output
+    gene_vec2rn<- convert_orthologs(
+        gene_df = rownames(exp_mouse_df)[seq(1, 100)],
+        input_species = "mouse",
+        gene_output = "rownames",
+        method = method
+    )
+    testthat::expect_true(methods::is(gene_vec2rn, "data.frame"))
+    testthat::expect_true(methods::is(rownames(gene_vec2rn), "character")) 
+    testthat::expect_true(methods::is(gene_vec2rn$input_gene, "vector"))
+    testthat::expect_true(any(rownames(gene_vec2rn)!=gene_vec2rn$input_gene))
+    
+    # vector input - dataframe column output
+    gene_vec2col<- convert_orthologs(
+        gene_df = rownames(exp_mouse_df)[seq(1, 100)],
+        input_species = "mouse",
+        gene_output = "columns",
+        method = method
+    )
+    testthat::expect_true(methods::is(gene_vec2col, "data.frame"))
+    testthat::expect_true(methods::is(rownames(gene_vec2col), "character")) 
+    testthat::expect_true(methods::is(gene_vec2col$input_gene, "vector"))
+    testthat::expect_true(methods::is(gene_vec2col$ortholog_gene, "vector"))
+    testthat::expect_true(
+        any(rownames(gene_vec2col)!=gene_vec2col$ortholog_gene))
+    testthat::expect_true(all(rownames(gene_vec2col)==gene_vec2col$input_gene))
 
     #### Aggregate #####
     # sparse matrix ==> sparse matrix: as rownames
@@ -120,7 +156,8 @@ test_that("convert_orthologs works", {
         gene_input = "rownames",
         gene_output = "rownames",
         non121_strategy = "sum",
-        as_sparse = TRUE
+        as_sparse = TRUE,
+        method = method
     )
     testthat::expect_equal(methods::is(gene_smat_sum, "sparseMatrix"), TRUE)
     testthat::expect_equal(has_gene_cols(gene_smat_sum), 0)
@@ -137,4 +174,16 @@ test_that("convert_orthologs works", {
     testthat::expect_equal(methods::is(gene_babel, "sparseMatrix"), TRUE)
     testthat::expect_equal(has_gene_cols(gene_babel), 0)
     testthat::expect_gte(nrow(gene_babel), 12000)
+    
+    #### same species 
+    gene_mat_same <- convert_orthologs(
+        gene_df = exp_mouse_smat,
+        input_species = "mouse", 
+        output_species = "mouse",
+        as_sparse = TRUE,
+        method = method
+    )
+    testthat::expect_equal(methods::is(gene_mat_same, "sparseMatrix"), TRUE)
+    testthat::expect_equal(has_gene_cols(gene_mat_same), 0)
+    testthat::expect_gte(nrow(gene_mat_same), 13000)
 })